Last refactoring for #1600, move UDT tests, rename classes (#1612)
