+++ testing/src/main/java/io/stargate/it/http/RestApiv2RowsTest.java
@@ -49,7 +49,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
@@ -73,7 +72,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @NotThreadSafe
 @ExtendWith(CqlSessionExtension.class)
 @CqlSessionSpec()
-public class RestApiv2Test extends BaseIntegrationTest {
+public class RestApiv2RowsTest extends BaseIntegrationTest {
   private String keyspaceName;
   private String tableName;
   private static String authToken;
@@ -88,17 +87,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
   private static final ObjectReader LIST_OF_MAPS_GETRESPONSE_READER =
       objectMapper.readerFor(ListOfMapsGetResponseWrapper.class);
 
-  private static final ObjectReader MAP_GETRESPONSE_READER =
-      objectMapper.readerFor(MapGetResponseWrapper.class);
-
   static class ListOfMapsGetResponseWrapper extends GetResponseWrapper<List<Map<String, Object>>> {
     public ListOfMapsGetResponseWrapper() {
-      super(-1, null, null);
-    }
-  }
-
-  static class MapGetResponseWrapper extends GetResponseWrapper<Map<String, Object>> {
-    public MapGetResponseWrapper() {
       super(-1, null, null);
     }
   }
@@ -151,8 +141,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithQuery() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -183,7 +173,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithQuery2Filters() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "age int", "firstName text"),
@@ -238,8 +228,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithQueryAndRaw() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -325,8 +315,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithNotFound() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -357,7 +347,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithInQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "firstName text"),
@@ -410,7 +400,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
   //   this behavior is what exists.
   @Test
   public void getRowsWithExistsQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "firstName text", "enabled boolean"),
@@ -439,7 +429,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithSetContainsQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "tags set<text>", "firstName text"),
@@ -487,7 +477,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithContainsKeyQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "attributes map<text,text>", "firstName text"),
@@ -535,7 +525,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithContainsEntryQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "attributes map<text,text>", "firstName text"),
@@ -581,7 +571,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithTimestampQuery() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "firstName text", "created timestamp"),
@@ -614,7 +604,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getAllRowsWithPaging() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "firstName text"),
@@ -679,7 +669,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getAllRowsNoPaging() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "firstName text"),
@@ -721,7 +711,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
         .as("Disabled because MVs are not enabled by default on a Cassandra 4 backend")
         .isFalse();
 
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     tableName = "tbl_mvread_" + System.currentTimeMillis();
     createTestTable(
         tableName,
@@ -774,8 +764,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getInvalidWhereClause() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -804,8 +794,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRows() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     // To try to ensure we actually find the right entry, create one other entry first
     Map<String, String> row = new HashMap<>();
@@ -852,7 +842,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
   // "Stringified" (non-JSON, CQL literal) notation.
   @Test
   public void getRowsWithTupleStringified() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "data tuple<int,boolean,text>", "alt_id uuid"),
@@ -877,7 +867,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
   // standard JSON payload
   @Test
   public void getRowsWithTupleTyped() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "data tuple<int,boolean,text>", "alt_id uuid"),
@@ -906,7 +896,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsWithUDT() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
 
     // create UDT: note -- UDT names must be lower-case it seems (mixed case fails)
     String udtString =
@@ -988,8 +978,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsNotFound() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1019,8 +1009,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void getRowsRaw() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1139,8 +1129,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void addRow() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1161,7 +1151,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void addRowWithCounter() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "counter counter"),
@@ -1182,7 +1172,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void addRowWithList() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("name text", "email list<text>"),
@@ -1215,8 +1205,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void addRowInvalidField() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1239,8 +1229,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void addRowWithInvalidJson() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     RestUtils.post(
         authToken,
@@ -1251,8 +1241,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void updateRow() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1282,8 +1272,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void updateRowRaw() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1336,7 +1326,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void updateRowWithCounter() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "counter counter"),
@@ -1400,7 +1390,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void updateRowWithMultipleCounters() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTable(
         tableName,
         Arrays.asList("id text", "counter1 counter", "counter2 counter"),
@@ -1443,8 +1433,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void updateRowWithInvalidJson() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1467,8 +1457,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void patchRow() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1512,8 +1502,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void patchRowRaw() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1560,8 +1550,8 @@ public class RestApiv2Test extends BaseIntegrationTest {
 
   @Test
   public void deleteRow() throws IOException {
-    createKeyspace(keyspaceName);
-    createTable(keyspaceName, tableName);
+    createTestKeyspace(keyspaceName);
+    createTestTable(keyspaceName, tableName);
 
     String rowIdentifier = UUID.randomUUID().toString();
     Map<String, String> row = new HashMap<>();
@@ -1756,334 +1746,13 @@ public class RestApiv2Test extends BaseIntegrationTest {
   /************************************************************************
    */
 
-  @Test
-  public void createUdt() throws IOException {
-    createKeyspace(keyspaceName);
-
-    // create UDT
-    String udtString =
-        "{\"name\": \"udt1\", \"fields\":"
-            + "[{\"name\":\"firstname\",\"typeDefinition\":\"text\"},"
-            + "{\"name\":\"birthdate\",\"typeDefinition\":\"date\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    // throws error because same name, but ifNotExists = false
-    String response =
-        RestUtils.post(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-            udtString,
-            HttpStatus.SC_BAD_REQUEST);
-
-    String typeName = "udt1";
-
-    ApiError apiError = objectMapper.readValue(response, ApiError.class);
-    assertThat(apiError.getCode()).isEqualTo(HttpStatus.SC_BAD_REQUEST);
-    assertThat(apiError.getDescription())
-        .isEqualTo(
-            String.format(
-                "Bad request: A type named \"%s\".%s already exists", keyspaceName, typeName));
-
-    // don't create and don't throw exception because ifNotExists = true
-    udtString =
-        "{\"name\": \"udt1\", \"ifNotExists\": true,"
-            + "\"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-  }
-
-  @Test
-  public void updateInvalidUdt() throws IOException {
-    createKeyspace(keyspaceName);
-
-    // create UDT
-    String udtString =
-        "{\"name\": \"udt1\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    // add existing field
-    udtString =
-        "{\"name\": \"udt1\", \"addFields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.put(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-
-    // missing add-type and rename-type
-    udtString =
-        "{\"name\": \"udt1\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.put(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-  }
-
-  @Test
-  public void updateUdt() throws IOException {
-    createKeyspace(keyspaceName);
-
-    // create UDT
-    String udtString =
-        "{\"name\": \"udt1\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    // update UDT: add new field
-    udtString =
-        "{\"name\": \"udt1\", \"addFields\":[{\"name\":\"lastname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.put(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_OK);
-
-    // udpate UDT: rename fields
-    udtString =
-        "{\"name\": \"udt1\",\"renameFields\":"
-            + "[{\"from\":\"firstname\",\"to\":\"name1\"}, {\"from\":\"lastname\",\"to\":\"name2\"}]}";
-
-    RestUtils.put(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_OK);
-
-    // retrieve UDT
-    String body =
-        RestUtils.get(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "udt1"),
-            HttpStatus.SC_OK);
-
-    MapGetResponseWrapper getResponseWrapper = MAP_GETRESPONSE_READER.readValue(body);
-    Map<String, Object> response = getResponseWrapper.getData();
-
-    assertThat(response.size()).isEqualTo(3);
-    assertThat(response.get("name")).isEqualTo("udt1");
-    List<Map<String, String>> fields = (List<Map<String, String>>) response.get("fields");
-    assertThat(fields.size()).isEqualTo(2);
-
-    Set<String> fieldNames = new HashSet<>();
-    fieldNames.add(fields.get(0).get("name"));
-    fieldNames.add(fields.get(1).get("name"));
-
-    assertThat(fieldNames.contains("name1")).isTrue();
-    assertThat(fieldNames.contains("name2")).isTrue();
-
-    // create UDT
-    udtString = "{\"name\": \"udt2\", \"fields\":[{\"name\":\"age\",\"typeDefinition\":\"int\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    // update UDT: add and rename field
-    udtString =
-        "{\"name\": \"udt2\","
-            + "\"addFields\":[{\"name\":\"name\",\"typeDefinition\":\"text\"}]},"
-            + "\"renameFields\": [{\"from\": \"name\", \"to\": \"firstname\"}";
-
-    RestUtils.put(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_OK);
-  }
-
-  @Test
-  public void testInvalidUdtType() throws IOException {
-    createKeyspace(keyspaceName);
-
-    String udtString =
-        "{\"name\": \"udt1\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"invalid_type\"}}]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-
-    udtString = "{\"name\": \"udt1\", \"fields\":[]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-
-    udtString = "{\"name\": \"udt1\", \"fields\":[{\"name\":\"firstname\"}}]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-
-    udtString = "{\"name\": \"udt1\", \"fields\":[{\"typeDefinition\":\"text\"}}]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_BAD_REQUEST);
-  }
-
-  @Test
-  public void dropUdt() throws IOException {
-    createKeyspace(keyspaceName);
-
-    String udtString =
-        "{\"name\": \"test_udt1\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    RestUtils.delete(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "test_udt1"),
-        HttpStatus.SC_NO_CONTENT);
-
-    // delete a non existent UDT
-    RestUtils.delete(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "test_udt1"),
-        HttpStatus.SC_BAD_REQUEST);
-
-    // delete an UDT in use
-    udtString =
-        "{\"name\": \"fullname\", \"fields\":"
-            + "[{\"name\":\"firstname\",\"typeDefinition\":\"text\"},"
-            + "{\"name\":\"lastname\",\"typeDefinition\":\"text\"}]}";
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    createTestTable(
-        tableName,
-        Arrays.asList("id text", "name fullname"),
-        Collections.singletonList("id"),
-        null);
-
-    String res =
-        RestUtils.delete(
-            authToken,
-            String.format(
-                "%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "fullname"),
-            HttpStatus.SC_BAD_REQUEST);
-  }
-
-  @Test
-  public void getUdt() throws IOException {
-    createKeyspace(keyspaceName);
-
-    String udtString =
-        "{\"name\": \"test_udt1\", \"fields\":[{\"name\":\"arrival\",\"typeDefinition\":\"timestamp\"}]}";
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-        udtString,
-        HttpStatus.SC_CREATED);
-
-    String body =
-        RestUtils.get(
-            authToken,
-            String.format(
-                "%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "test_udt1"),
-            HttpStatus.SC_OK);
-
-    MapGetResponseWrapper getResponseWrapper = MAP_GETRESPONSE_READER.readValue(body);
-    Map<String, Object> response = getResponseWrapper.getData();
-
-    assertThat(response.size()).isEqualTo(3);
-    assertThat(response.get("name")).isEqualTo("test_udt1");
-    List<Map<String, String>> fields = (List<Map<String, String>>) response.get("fields");
-    assertThat(fields.size()).isEqualTo(1);
-    assertThat(fields.get(0).get("name")).isEqualTo("arrival");
-    assertThat(fields.get(0).get("typeDefinition")).isEqualTo("timestamp");
-
-    // get non existent UDT
-    RestUtils.get(
-        authToken,
-        String.format(
-            "%s:8082/v2/schemas/keyspaces/%s/types/%s", host, keyspaceName, "invalid_udt"),
-        HttpStatus.SC_NOT_FOUND);
-  }
-
-  @Test
-  public void listAllTypes() throws IOException {
-    createKeyspace(keyspaceName);
-
-    String body =
-        RestUtils.get(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-            HttpStatus.SC_OK);
-
-    ListOfMapsGetResponseWrapper getResponseWrapper =
-        LIST_OF_MAPS_GETRESPONSE_READER.readValue(body);
-    List<Map<String, Object>> response = getResponseWrapper.getData();
-    assertThat(response.size()).isEqualTo(0);
-
-    // creates 10 UDTs
-    String udtString =
-        "{\"name\": \"%s\", \"fields\":[{\"name\":\"firstname\",\"typeDefinition\":\"text\"}]}";
-    for (int i = 0; i < 10; i++) {
-      RestUtils.post(
-          authToken,
-          String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-          String.format(udtString, "udt" + i),
-          HttpStatus.SC_CREATED);
-    }
-
-    body =
-        RestUtils.get(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/types", host, keyspaceName),
-            HttpStatus.SC_OK);
-
-    getResponseWrapper = LIST_OF_MAPS_GETRESPONSE_READER.readValue(body);
-    response = getResponseWrapper.getData();
-    assertThat(response.size()).isEqualTo(10);
-
-    List<Map<String, String>> fields = (List<Map<String, String>>) response.get(0).get("fields");
-    assertThat(fields.size()).isEqualTo(1);
-    assertThat(fields.get(0).get("name")).isEqualTo("firstname");
-    assertThat(fields.get(0).get("typeDefinition")).isEqualTo("text");
-  }
-
   /*
   /************************************************************************
   /* Helper methods for setting up tests
   /************************************************************************
    */
 
-  private void createTable(String keyspaceName, String tableName) throws IOException {
+  private void createTestTable(String keyspaceName, String tableName) throws IOException {
     TableAdd tableAdd = new TableAdd();
     tableAdd.setName(tableName);
 
@@ -2193,7 +1862,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
         HttpStatus.SC_CREATED);
   }
 
-  private void createKeyspace(String keyspaceName) throws IOException {
+  private void createTestKeyspace(String keyspaceName) throws IOException {
     String createKeyspaceRequest =
         String.format("{\"name\": \"%s\", \"replicas\": 1}", keyspaceName);
 
@@ -2273,7 +1942,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
   }
 
   private String setupClusteringTestCase() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTableWithClustering(keyspaceName, tableName);
 
     String rowIdentifier = "1";
@@ -2325,7 +1994,7 @@ public class RestApiv2Test extends BaseIntegrationTest {
   }
 
   private void setupMixedClusteringTestCase() throws IOException {
-    createKeyspace(keyspaceName);
+    createTestKeyspace(keyspaceName);
     createTestTableWithMixedClustering(keyspaceName, tableName);
 
     Map<String, String> row = new HashMap<>();
